requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==44.0.0
databases==0.9.0
dateparser==1.2.0
fastapi==0.115.6
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Mako==1.3.8
MarkupSafe==3.0.2
multidict==6.1.0
numpy==2.2.0
pandas==2.2.3
propcache==0.2.1
psycopg2-binary==2.9.10
pycoingecko==3.2.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.3
pydantic_core==2.27.1
python-binance==1.0.25
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
uvicorn==0.32.1
websockets==14.1
yarl==1.18.3
